requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clr==1.0.3
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-WTF==0.14.3
fonttools==4.13.0
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
memory-profiler==0.57.0
networkx==2.4
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
psutil==5.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.24.0
scikit-image==0.17.2
scipy==1.5.1
selenium==3.141.0
Shapely==1.7.0
six==1.15.0
soupsieve==2.0.1
svgwrite==1.4
tifffile==2020.7.4
Tree==0.2.4
urllib3==1.25.9
validators==0.18.0
webdriver-manager==3.2.2
Werkzeug==1.0.1
WTForms==2.3.3